refactor(slop): remove more unncessary files
